Make srclists into a module to allow setup.py installation and tests to work generally.
py.test works if the parent directory is in the PYTHONPATH (e.g. PYTHONPATH="."
py.test)

This commit unfortunately alters all lines in srclist_by_beam.py. This was to
make a main function that the tests can call, to isolate the command-line
arguments into a function that yields defaults, and so the tests can run
srclist_by_beam with custom command-line arguments.

I can't work out how to make the source list files actually install if using
"pip install ." or "python setup.py install", but maybe that's not a good idea anyway?